+++ packages/tuya-panel-health-sdk/src/components/dateSwitch/index.tsx
@@ -1,0 +1,5 @@
+import DaySwitch from './DaySwitch';
+import WeekSwitch from './WeekSwitch';
+import MonthSwitch from './MonthSwitch';
+
+export { DaySwitch, WeekSwitch, MonthSwitch };