+++ packages/tuya-panel-lamp-sdk/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.2](https://github.com/tuya/tuya-panel-sdk/compare/@tuya/tuya-panel-lamp-sdk@1.9.1...@tuya/tuya-panel-lamp-sdk@1.9.2) (2022-01-13)
+
+**Note:** Version bump only for package @tuya/tuya-panel-lamp-sdk
+
+
+
+
+
+## [1.9.1](https://github.com/tuya/tuya-panel-sdk/compare/@tuya/tuya-panel-lamp-sdk@1.9.0...@tuya/tuya-panel-lamp-sdk@1.9.1) (2022-01-11)
+
+**Note:** Version bump only for package @tuya/tuya-panel-lamp-sdk
+
+
+
+
+
 # [1.9.0](https://github.com/tuya/tuya-panel-sdk/compare/@tuya/tuya-panel-lamp-sdk@1.8.0...@tuya/tuya-panel-lamp-sdk@1.9.0) (2021-12-30)